L-101. Lift state up. Move userInput state, handleChange to App so that userInput can be accessible to Results. Refactor UserInput, Result for this.
