+++ src/components/Results.jsx
@@ -1,19 +1,16 @@
 /**
  * @returns Results component
  */
-export default function Results({ investment }) {
-    // Calculate table data
-    const tableDataValues = calculateInvestmentResults(deriveInvestmentValues(investment));
-
+export default function Results({ input }) {
     return (
         <table id="result">
             <thead>
                 <tr>
-                   
+
                 </tr>
             </thead>
             <tbody>
-               
+
             </tbody>
         </table>
     );
