+++ src/components/UserInput.jsx
@@ -1,33 +1,7 @@
-import { useState } from "react";
-
 /**
  * @returns UserInput component
  */
-export default function UserInput() {
-
-    const [userInput, setUserInput] = useState({
-        initialInvestment: 10000,
-        annualInvestment: 1200,
-        expectedReturn: 6,
-        duration: 10
-    });
-
-    /**
-     * Handles input changes.
-     * Notice the event parameter for the custom onChange function in input.
-     * 
-     * @param {*} inputIdentifier 
-     * @param {*} newValue 
-     */
-    function handleChange(inputIdentifier, newValue) {
-        setUserInput((previousUserInput) => {
-            return {
-                ...previousUserInput,
-                [inputIdentifier]: newValue
-            };
-        });
-    }
-
+export default function UserInput({ userInput, onChange }) {
     return (
         <section id="user-input">
             <div className="input-group">
@@ -36,7 +10,7 @@ export default function UserInput() {
                     <input
                         type="number" required
                         value={userInput.initialInvestment}
-                        onChange={(event) => handleChange("initialInvestment", event.target.value)}
+                        onChange={(event) => onChange("initialInvestment", event.target.value)}
                     />
                 </p>
                 <p>
@@ -44,7 +18,7 @@ export default function UserInput() {
                     <input
                         type="number" required
                         value={userInput.annualInvestment}
-                        onChange={(event) => handleChange("annualInvestment", event.target.value)}
+                        onChange={(event) => onChange("annualInvestment", event.target.value)}
                     />
                 </p>
             </div>
@@ -54,7 +28,7 @@ export default function UserInput() {
                     <input
                         type="number" required
                         value={userInput.expectedReturn}
-                        onChange={(event) => handleChange("expectedReturn", event.target.value)}
+                        onChange={(event) => onChange("expectedReturn", event.target.value)}
                     />
                 </p>
                 <p>
@@ -62,7 +36,7 @@ export default function UserInput() {
                     <input
                         type="number" required
                         value={userInput.duration}
-                        onChange={(event) => handleChange("duration", event.target.value)}
+                        onChange={(event) => onChange("duration", event.target.value)}
                     />
                 </p>
             </div>
